Adds epsilon-first strategy
Adds a function to generate the epsilon-first strategy.

Also adds the `SimulationParameters` type and passes an instance of
that type to the `getAllocation` function passed to the
`RunSimulation` function. The `SimulationParameters` type generalizes
the `numLevers` parameter (now `NumLevers`) and includes `NumRounds`,
which is needed for the epsilon-first strategy, there is also room to
add other parameters in the future.