Refactors `EvaluateActions` handling
Refactors the handling of `EvaluateActions` in all agents (other than
the VDBE agent). These agents were previously using the utility function
`GetActionSampleAverages`, which loops over all of the historical data
to calculate sample average values for each action. This updates the
agents to maintain `ActionValueEstimates` and update values on each
step, which makes the computation time for `EvaluateActions` constant as
time goes on, rather than linear (roughly).

Creates the function `UpdateActionSampleAverages` for handling this
logic. Refactors the VDBE agent to use this same utility function, which
was based on the VDBE's `EvaluateActions` method.